Refactor Acme plugin into AcmeChallenge objects, http01
